Convert Operator Jobs#Show to Typescript
