+++ operator_ui/@types/use-react-hooks.d.ts
@@ -1,0 +1,1 @@
+declare module 'use-react-hooks'
